Move requantize.py to the quatnization/ dir.
